rename LikedContentList to LikedContentFeedConnected
