+++ packages/apollos-ui-connected/src/LikedContentFeedConnected/index.js
@@ -1,0 +1,2 @@
+export LikedContentFeedConnected from './LikedContentFeedConnected';
+export GET_LIKED_CONTENT from './getLikedContent';
