+++ packages/apollos-ui-connected/src/index.js
@@ -1,3 +1,5 @@
 export { LiveProvider, LiveConsumer } from './live';
-export LikedContentList, { GET_LIKED_CONTENT } from './LikedContentList';
+export LikedContentFeedConnected, {
+  GET_LIKED_CONTENT,
+} from './LikedContentFeedConnected';
 export { fetchMoreResolver } from './utils';
